Move django-unrelated stuff into evap.tools
