+++ evap/evaluation/tools.py
@@ -1,7 +1,6 @@
 import datetime
 import typing
 from abc import ABC, abstractmethod
-from collections import defaultdict
 from collections.abc import Iterable, Mapping
 from typing import TYPE_CHECKING, Any, TypeVar
 from urllib.parse import quote
@@ -17,6 +16,8 @@ from django.utils.datastructures import MultiValueDict
 from django.utils.translation import get_language
 from django.views.generic import FormView
 
+from evap.tools import date_to_datetime
+
 if TYPE_CHECKING:
     from django_stubs_ext import StrOrPromise  # use proper definition with mypy
 else:
@@ -27,8 +28,6 @@ else:
 
 M = TypeVar("M", bound=Model)
 T = TypeVar("T")
-Key = TypeVar("Key")
-Value = TypeVar("Value")
 CellValue = str | int | float | None
 CV = TypeVar("CV", bound=CellValue)
 
@@ -39,20 +38,6 @@ def openid_login_is_active() -> bool:
 
 def password_login_is_active() -> bool:
     return not openid_login_is_active()
-
-
-def unordered_groupby(key_value_pairs: Iterable[tuple[Key, Value]]) -> dict[Key, list[Value]]:
-    """
-    We need this in several places: Take list of (key, value) pairs and make
-    them into the aggregated all-values-of-every-unique-key dict. Note that
-    this slightly differs from itertools.groupby (and uniq), as we don't
-    require anything to be sorted and you get a dict as return value.
-    """
-    result = defaultdict(list)
-    for key, value in key_value_pairs:
-        result[key].append(value)
-
-    return dict(result)
 
 
 def get_object_from_dict_pk_entry_or_logged_40x(
@@ -112,13 +97,16 @@ def sort_formset(request: HttpRequest, formset: BaseFormSet) -> None:
         formset.forms.sort(key=lambda f: f.cleaned_data.get("order", 9001))
 
 
-def date_to_datetime(date: datetime.date) -> datetime.datetime:
-    return datetime.datetime(year=date.year, month=date.month, day=date.day)
-
-
 def vote_end_datetime(vote_end_date: datetime.date) -> datetime.datetime:
     # The evaluation actually ends at EVALUATION_END_OFFSET_HOURS:00 of the day AFTER self.vote_end_date.
     return date_to_datetime(vote_end_date) + datetime.timedelta(hours=24 + settings.EVALUATION_END_OFFSET_HOURS)
+
+
+def translate(**kwargs):
+    # pylint is really buggy with this method.
+    # pylint: disable=unused-variable, useless-suppression
+    # get_language may return None if there is no session (e.g. during management commands)
+    return property(lambda self: getattr(self, kwargs[get_language() or "en"]))
 
 
 def get_parameter_from_url_or_session(request: HttpRequest, parameter: str, default=False) -> bool:
@@ -129,13 +117,6 @@ def get_parameter_from_url_or_session(request: HttpRequest, parameter: str, defa
         result = {"true": True, "false": False}.get(result_str.lower())  # convert parameter to boolean
     request.session[parameter] = result  # store value for session
     return result
-
-
-def translate(**kwargs):
-    # pylint is really buggy with this method.
-    # pylint: disable=unused-variable, useless-suppression
-    # get_language may return None if there is no session (e.g. during management commands)
-    return property(lambda self: getattr(self, kwargs[get_language() or "en"]))
 
 
 EmailT = TypeVar("EmailT", str, None)
@@ -155,15 +136,6 @@ def clean_email(email: EmailT) -> EmailT:
 def capitalize_first(string: StrOrPromise) -> str:
     """Realize lazy promise objects and capitalize first letter."""
     return string[0].upper() + string[1:]
-
-
-def ilen(iterable: Iterable) -> int:
-    return sum(1 for _ in iterable)
-
-
-def assert_not_none(value: T | None) -> T:
-    assert value is not None
-    return value
 
 
 class FormsetView(FormView):
