+++ evap/results/tools.py
@@ -20,7 +20,8 @@ from evap.evaluation.models import (
     TextAnswer,
     UserProfile,
 )
-from evap.evaluation.tools import discard_cached_related_objects, unordered_groupby
+from evap.evaluation.tools import discard_cached_related_objects
+from evap.tools import unordered_groupby
 
 STATES_WITH_RESULTS_CACHING = {Evaluation.State.EVALUATED, Evaluation.State.REVIEWED, Evaluation.State.PUBLISHED}
 STATES_WITH_RESULT_TEMPLATE_CACHING = {Evaluation.State.PUBLISHED}
