+++ evap/staff/forms.py
@@ -31,11 +31,12 @@ from evap.evaluation.models import (
     TextAnswer,
     UserProfile,
 )
-from evap.evaluation.tools import clean_email, date_to_datetime
+from evap.evaluation.tools import clean_email
 from evap.results.tools import STATES_WITH_RESULT_TEMPLATE_CACHING, STATES_WITH_RESULTS_CACHING, cache_results
 from evap.results.views import update_template_cache, update_template_cache_of_published_evaluations_in_course
 from evap.staff.tools import remove_user_from_represented_and_ccing_users
 from evap.student.models import TextAnswerWarning
+from evap.tools import date_to_datetime
 
 logger = logging.getLogger(__name__)
